[logistic_plugin] fix error on comparison with size_type

+++ src/libraries/logistic_plugin/LogisticPlugin.cpp
@@ -121,7 +121,7 @@ void LogisticPlugin::Receive( const sword::SimToClient& message, const bg::date&
 int LogisticPlugin::GetConsignCount( E_LogisticType eLogisticType ) const
 {
     int resolverIndex = static_cast< int >( eLogisticType );
-    if( resolverIndex < resolvers_.size() )
+    if( resolverIndex < static_cast< int >( resolvers_.size() ) )
     {
         const ConsignResolver_ABC* pResolver = resolvers_[ resolverIndex ];
         if( pResolver )
